+++ src/App.js
@@ -1,36 +1,13 @@
-import { useState } from "react";
-
-import styled, { ThemeProvider } from "styled-components";
+import styled from "styled-components";
 import GlobalStyles from "./global-styles";
-
-const BaseTheme = {
-  color: '#222',
-  background: '#fff',
-};
-
-const DarkTheme = {
-  color: '#fff',
-  background: '#222',
-}
-
-const Container = styled.div`
-  padding: 2rem;
-  color: ${(props) => props.theme.color};
-  background: ${(props) => props.theme.background};
-`
+import Loading from "./components/Loading";
 
 function App() {
-  const [baseTheme, setTheme] = useState(true);
-  const toggleTheme = () => setTheme(!baseTheme);
-
   return (
-    <ThemeProvider theme={baseTheme ? BaseTheme : DarkTheme}>
+    <div styles={{padding: "2rem"}}>
       <GlobalStyles />
-      <Container>
-        <h1>Hello World</h1>
-        <button className="btn" onClick={toggleTheme}>toggle me</button>
-      </Container>
-    </ThemeProvider>
+      <Loading />
+    </div>
   );
 }
